Refactor calendar page (#445)
* Refactor calendar page

This split out the Calendar page into more manageable pieces.

Breaks out the Header, Footer and NoJS calendar page.
